use shared_ptr to avoid lifetime issues
When running multiple async operations on the same object, where one
of the operations can destroy the object, leads to use-after-free.
This could have happened when the client object finished sending a
message after it was removed from the room.

The standard way to avoid this seems to be to just use shared_ptr.